requirements.txt
py3exiv2 
pyexiv2
google-generativeai